+++ tests/default.nix
@@ -32,6 +32,7 @@ let
     // prefix "secureSocketUrl" (tryImport "secure-socket-url.nix")
     // prefix "url" (tryImport "url.nix")
     // prefix "urlHost" (tryImport "url-host.nix")
+    // prefix "authority" (tryImport "authority.nix")
     // prefix "ipListener" (tryImport "ip-listener.nix")
     // prefix "listener" (tryImport "listener.nix")
     // prefix "transport" (tryImport "transport.nix")